change price type in db

+++ database/migrations/2024_03_04_073856_create_sample_details_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('sample_id')->references('id')->on('samples');
             $table->string('description');
-            $table->integer('price');
+            $table->decimal('price', 3, 2);
             $table->string('duration')->nullable();
             $table->timestamps();
         });
